Fix linker errors: enum-to-int cast and add boost_regex
- connection.h: Revert to enum for timeout constants (avoids ODR
  linker issues with static const int), cast to int at call sites
  to satisfy boost::posix_time::seconds.
- configure.ac: Add boost_regex library check, needed by tools.cpp.

+++ connection.cpp
@@ -318,7 +318,7 @@ void Connection::accept()
 	try
 	{
 		++m_pendingRead;
-		m_readTimer.expires_from_now(boost::posix_time::seconds(Connection::readTimeout));
+		m_readTimer.expires_from_now(boost::posix_time::seconds((int)Connection::readTimeout));
 		m_readTimer.async_wait(boost::bind(&Connection::handleReadTimeout,
 			boost::weak_ptr<Connection>(shared_from_this()), boost::asio::placeholders::error));
 
@@ -358,7 +358,7 @@ void Connection::parseHeader(const boost::system::error_code& error)
 	try
 	{
 		++m_pendingRead;
-		m_readTimer.expires_from_now(boost::posix_time::seconds(Connection::readTimeout));
+		m_readTimer.expires_from_now(boost::posix_time::seconds((int)Connection::readTimeout));
 		m_readTimer.async_wait(boost::bind(&Connection::handleReadTimeout,
 			boost::weak_ptr<Connection>(shared_from_this()), boost::asio::placeholders::error));
 
@@ -448,7 +448,7 @@ void Connection::parsePacket(const boost::system::error_code& error)
 	try
 	{
 		++m_pendingRead;
-		m_readTimer.expires_from_now(boost::posix_time::seconds(Connection::readTimeout));
+		m_readTimer.expires_from_now(boost::posix_time::seconds((int)Connection::readTimeout));
 		m_readTimer.async_wait(boost::bind(&Connection::handleReadTimeout,
 			boost::weak_ptr<Connection>(shared_from_this()), boost::asio::placeholders::error));
 
@@ -516,7 +516,7 @@ void Connection::internalSend(OutputMessage_ptr msg)
 	try
 	{
 		++m_pendingWrite;
-		m_writeTimer.expires_from_now(boost::posix_time::seconds(Connection::writeTimeout));
+		m_writeTimer.expires_from_now(boost::posix_time::seconds((int)Connection::writeTimeout));
 		m_writeTimer.async_wait(boost::bind(&Connection::handleWriteTimeout,
 			boost::weak_ptr<Connection>(shared_from_this()), boost::asio::placeholders::error));
